refactor(freshell-sessions): split codex_locator tests into sibling file
Pure code-move to the repo's #[path]-included child-module convention
(pane_ledger_tests.rs et al.); both files now respect the <=1K-line
limit and the tests keep private-item access. 34 tests before and after.
